Move Languge Server init
This moves the language server and client init to a seperate file for
better extension.ts houskeeping
